Refactor ToolsModal to use ToolService for fetching tools
